% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bash_history.R
\docType{data}
\name{bash_history}
\alias{bash_history}
\title{Historia ostatnio wykonywanych poleceń.}
\format{Wektor z 19913 elementami.}
\usage{
data(bash_history)
}
\description{
Wektor ostatnio wykonywanych poleceń.
Każdy element wektora składa się z nazwy polecenia i argumentów polecenia,
rozdzielonych spacją.
}
\keyword{datasets}

