% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees.R
\docType{data}
\name{employees}
\alias{employees}
\title{The database with employees of Elka Faculty of University of Technology.}
\format{a data frame with 541 rows and three columns.}
\usage{
data(employees)
}
\description{
The dataset describing names, surnames and faculty employees' logins.
The subsequent columns in this dataset describe:
\itemize{
  \item name. The name of an employee.
  \item surname. The surname of an employee.
  \item login. The login of an employee on the Proton server.
}
}
\keyword{datasets}

