% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logowania.R
\docType{data}
\name{logowania}
\alias{logowania}
\title{Historia logowań do serwera Proton}
\format{Ramka danych z 59366 wierszami i 3 kolumnami.}
\usage{
data(logowania)
}
\description{
Zbiór danych opisujących historie logowania kto, skąd i kiedy logował się na Protona.
Kolejne kolumny w tym zbiorze danych opisują:
}
\details{
\itemize{
  \item login. Login użytkownika, który loguje się do Protona.
  \item host. Adres IP komputera, z którego dochodziło do logowania do Protona.
  \item data. Data logowania do Protona. Wiersze są posortowane po tej kolumnie.
}
}
\keyword{datasets}

