% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proton.R
\name{proton}
\alias{proton}
\title{,,Włam do Protona''}
\usage{
proton(...)
}
\arguments{
\item{...}{Funkcja proton przyjmuje rózne argumenty w zależności od tego, którą zagadkę rozwiązujemy.
Zobacz opis szczegółowy aby poznać pełną listę argumentów.}
}
\description{
Funkcja \code{proton} pozwala na rozwiązywanie zadań w grze ,,Włam do Protona''.
Wystarczy rozwiazać cztery zagadki by dostać się na konto Pietraszki!
}
\details{
W każdej sytuacji, jeżeli potrzebujemy dodatkowych wskazówek, należy dodać argument `wskazowka=TRUE`.

Aby dowiedzić się więcej o użtkowniku na serwerze proton należy podać argumenty `action = "login", login="XYZ".

Aby zalogować się do protona należy podać argumenty `action = "login", login="XYZ", password="ABC"`.
Jeżeli hasło pasuje do loginu, wtedy uzyskamy komunikat o poprawnym zalogowaniu.

Aby zalogowac się do innego serwera niż proton, należy podać argumenty `action = "server", host="XYZ"`.

Uwaga:

Gra ,,Włam do Protona'' jest projektem edukacyjnym Fundacji SmarterPoland dostępnym bezpłatnie.
Pomóż nam tworzyć kolejne gry edukacyjne przekazując darowiznę na konto fundacji wskazanym na stronie
\url{http://smarterpoland.pl/index.php/fundacja/}.
Nawet kwoty rzędu 10 zł uławtią nam tworzenie i utrzymywanie kolejnych projektów edukacyjnych!
Dziękujemy!
}
\examples{
\dontrun{
proton()
proton(wskazowka=TRUE)
}
}
\author{
Przemysław Biecek, \email{przemyslaw.biecek@gmail.com}, Fundacja SmarterPoland.pl
}

