% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top1000passwords.R
\docType{data}
\name{top1000passwords}
\alias{top1000passwords}
\title{Wektor 1000 najpopularniejszych haseł.}
\format{Wektor z 1000 elementami.}
\usage{
data(top1000passwords)
}
\description{
Wektor 1000 najczęściej stosowanych haseł.
Posortowana po częstości użycia, pierwsze hasła są często używane.
}
\keyword{datasets}

