% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bash_history.R
\docType{data}
\name{bash_history}
\alias{bash_history}
\title{The history of recently executed commands.}
\format{a character vector with 19913 elements.}
\usage{
data(bash_history)
}
\description{
The character vector of recently executed commands.
Each element of the vector consists of command's name and command's arguments
separated with a space.
}
\keyword{datasets}

