% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proton.R
\name{proton}
\alias{proton}
\title{,,Hack into Proton''}
\usage{
proton(...)
}
\arguments{
\item{...}{\code{proton} function is called by different arguments, which vary depending
on a riddle that we try to solve. See \code{Details} in order to get to know how the
whole list of arguments looks like.}
}
\description{
\code{proton} function enables a player to solve problems in ,,Hack into Proton'' game.
It is sufficient to solve four riddles in order to get into Pietraszko's account!
}
\details{
Every time when some additional hints are needed one should pass 
\code{hint=TRUE} argument to the \code{proton} function.

In order to get some more information about a user on the Proton server  
one should pass \code{action = "login"}, \code{login="XYZ"} arguments 
to the \code{proton} function.

In order to log into the Proton server one should pass \code{action = "login"},
\code{login="XYZ"}, \code{password="ABC"} arguments to the \code{proton} function.
If the password matches login, then one will receive a message about correct log.

In order to log into a server different from Proton one should pass 
\code{action = "server"}, \code{host="XYZ"} arguments to the \code{proton} function.

\strong{Attention}:

,,Hack into Proton'' game is free of charge, educational project of the SmarterPoland.pl Foundation.
By transfering a donation on the foundation's account which is 
shown on \url{http://smarterpoland.pl/index.php/fundacja/} site, you will help
us to create another educational games. Even a sum of 10 $ will faciliate
creating and maintaining subsequent educational projects!
Thank you!
}
\examples{
\dontrun{
proton()
proton(hint=TRUE)
}
}
\author{
Przemysław Biecek, \email{przemyslaw.biecek@gmail.com}, SmarterPoland.pl Foundation.
}

