% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proton.R
\name{proton}
\alias{proton}
\title{Infiltration of the Proton server}
\usage{
proton(...)
}
\arguments{
\item{...}{\code{proton} function is called by different arguments, which vary depending
on a problem that Bit is trying to solve. See \code{Details} in order to learn more about the list of possible arguments.}
}
\description{
The \code{proton} function is used for solving problems in the data-based game ,,Infiltration of the proton server''.
Solve four data-based puzzles in order to crack into the Pietraszko's account!
}
\details{
Every time when some additional hints are needed one should add
\code{hint=TRUE} argument to the \code{proton} function.

In order to get more information about a user on the Proton server
one should pass \code{action = "login"}, \code{login="XYZ"} arguments
to the \code{proton} function.

In order to log into the Proton server one should pass \code{action = "login"},
\code{login="XYZ"}, \code{password="ABC"} arguments to the \code{proton} function.
If the password matches login, then one will receive a message about successful login.

In order to log into a server different from Proton one should pass
\code{action = "server"}, \code{host="XYZ"} arguments to the \code{proton} function.

The ,,Infiltration of the Proton server'' game is a free of charge, educational project of the SmarterPoland.pl Foundation.
By transfering a donation on the foundation's account which is
shown on \url{http://smarterpoland.pl/index.php/fundacja/} site, you will help
us to create another educational games. Even a sum of 10 $ will faciliate
creating and maintaining subsequent educational projects!
Thank you!
}
\examples{
\dontrun{
proton()
proton(hint=TRUE)
}
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}, SmarterPoland.pl Foundation.
}

